Wielded items are again shown on the game screen.

+++ Main/Include/nonhuman.h
@@ -183,7 +183,7 @@ class CHARACTER
   dolphin,
   nonhumanoid,
  protected:
-  virtual uchar GetSpecialBodyPartFlags(ushort, bool = false) const;
+  virtual ushort GetSpecialBodyPartFlags(ushort, bool = false) const;
   virtual void SpecialTurnHandler() { UpdatePictures(); }
 );
